% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_1d}
\alias{data_1d}
\title{Dataset for 1dim domain data}
\format{
A dataset with 200 rows and 100 cols:
\describe{
\item{x}{a numeric vector (1:10)}
\item{y}{a numeric vector with random values}
}
}
\usage{
data(data_1d)
}
\description{
A dataset with 200 rows (discrete evaluation of the functional data) and 100 cols (time instants).
}
\examples{
data(data_1d)
head(data_1d)
}
\keyword{datasets}
