% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2d}
\alias{data_2d}
\title{Dataset for 2dim domain data}
\format{
A list with 20 10x10 matrices:
\describe{
  \item{x}{a numeric vector (1:10)}
  \item{y}{a numeric vector with random values}
}
}
\usage{
data(data_2d)
}
\description{
A list with 20 10x10 matrices: each matrix are the evaluation of the surface in a specific time instant (number of element of the list).
}
\examples{
data(data_2d)
head(data_2d)
}
\keyword{datasets}
