% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC_KO_Rinterface.R
\name{KO_check_hps_2d}
\alias{KO_check_hps_2d}
\title{KO_check_hps_2d}
\arguments{
\item{X}{\strong{\verb{numeric matrix}}. Each row (m) represents a point of the domain in which the surface is evaluated.
The surface is represented by a grid, of dimensions (m1,m2), such that their product is m.
Then, the grid is encapsulated into a vector, column after column. Each column (n) represents a time instant.
Some auxiliary functions (\link{data_2d_wrapper_from_list}, \link{data_2d_wrapper_from_array}) are available for wrapping data into a coherent data structure for the algorithm.}

\item{dim_x1}{\strong{\code{integer}}. The number of discrete evaluations along dimension one (has to be m1).}

\item{dim_x2}{\strong{\code{integer}}. The number of discrete evaluations along dimension one (has to be m2).}
}
\value{
\strong{\code{list}} whose items are:
\itemize{
\item 'P-values ADF': matrix containing the pointwise ADF test p-values.
}
}
\description{
Evaluate the pointwise Augmented DIckey-Fuller (ADF) test p-values for the available evaluations of the surface. Implemented as in \code{tseries} CRAN package.
}
\references{
\itemize{
\item Source code: \href{https://github.com/AndreaEnricoFranzoni/PPCforAutoregressiveOperator}{PPCKO implementation}
\item Original implementation: \href{https://cran.r-project.org/web/packages/tseries/index.html}{tseries}
}
}
\author{
Andrea Enrico Franzoni
}
