% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC_KO_Rinterface.R
\name{KO_show_results}
\alias{KO_show_results}
\title{KO_show_results}
\arguments{
\item{results_ko}{\strong{\code{list}} as output from \link{PPC_KO}.}

\item{hp_ko}{\strong{\code{list}} (default: \strong{\code{NULL}}) as output from \link{KO_check_hps}. If NULL: p-values not plotted.}

\item{x_lab}{\strong{\code{string}} (default: \strong{\code{"x"}}). Name of the x-axis in the plots.}

\item{y_lab}{\strong{\code{string}} (default: \strong{\code{"y"}}). Name of the y-axis in the plots.}

\item{true_alphas}{\strong{\code{bool}} (default: \strong{\code{FALSE}}). If plotting validation errors, if FALSE, the values of alpha are put equally spaced. If TRUE, as their real value}
}
\value{
No return. It simply prints and plots.
}
\description{
Print on the R console regularization parameter used, number of PPCs retained and cumulative explanatory power of the PPCs retained.
Plot pointwise p-value (eventually), prediction at the next instant comparing to the last one available,
the validation errors (eventually, if saved), direction and weight for each PPC retained, direction and weight of each PPC retained
as mean perturbation (mean function, mean function added/subtracted to PPCs direction and weight multiplied by their scores standard deviation).
}
\references{
Source code: \href{https://github.com/AndreaEnricoFranzoni/PPCforAutoregressiveOperator}{PPCKO implementation}
}
\seealso{
\link{PPC_KO}, \link{KO_check_hps}
}
\author{
Andrea Enrico Franzoni
}
