% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC_KO_Rinterface.R
\name{KO_show_results_2d}
\alias{KO_show_results_2d}
\title{KO_show_results_2d}
\arguments{
\item{results_ko}{\strong{\code{list}} as output from \link{PPC_KO_2d}.}

\item{hp_ko}{\strong{\code{list}} (default: \strong{\code{NULL}}) as output from \link{KO_check_hps_2d}. If NULL: p-values not plotted.}

\item{x1_lab}{\strong{\code{string}} (default: \strong{\code{"x1"}}). Name of the ax of the first dimension in the plots.}

\item{x2_lab}{\strong{\code{string}} (default: \strong{\code{"x2"}}). Name of the ax of the second dimension in the plots.}

\item{z_lab}{\strong{\code{string}} (default: \strong{\code{"value"}}). Name of the z-ax in the plots.}

\item{true_alphas}{\strong{\code{bool}} (default: \strong{\code{FALSE}}). If plotting validation errors, if FALSE, the values of alpha are put equally spaced. If TRUE, as their real value}
}
\value{
No return. It simply prints and plots.
}
\description{
Print on the R console regularization parameter used, number of PPCs retained and cumulative explanatory power of the PPCs retained.
Plot pointwise p-value (eventually), prediction at the next instant comparing to the last one available, the increment,
the validation errors (eventually, if saved), direction and weight for each PPC retained.
}
\references{
Source code: \href{https://github.com/AndreaEnricoFranzoni/PPCforAutoregressiveOperator}{PPCKO implementation}
}
\seealso{
\link{PPC_KO_2d}, \link{KO_check_hps_2d}
}
\author{
Andrea Enrico Franzoni
}
