% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_1d}
\alias{data_1d}
\title{Functional time series of curves}
\format{
A numeric matrix with 200 rows and 100 cols.
}
\usage{
data(data_1d)
}
\description{
FAR(1) process generating data.
Every row (200) represents one point of the domain (0,1) for which the curve evaluation is available.
Every column (100) represents the curve at one time instant.
}
\examples{
data(data_1d)
head(data_1d)
}
\keyword{datasets}
