% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2d}
\alias{data_2d}
\title{Functional time series of surfaces}
\format{
A list with 20 10x10 matrices.
}
\usage{
data(data_2d)
}
\description{
FAR(1) process generating data.
Every element of the list represents the surface at one time instant.
Every element of the list is a numeric matrix: it represents the grid over which the
evaluations of the surface are available.
}
\examples{
data(data_2d)
head(data_2d)
}
\keyword{datasets}
