% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC_KO_Rinterface.R
\name{data_2d_wrapper_from_array}
\alias{data_2d_wrapper_from_array}
\title{data_2d_wrapper_from_array}
\arguments{
\item{Xt}{\strong{\verb{numeric vector}}, dimensions (dim_x1,dim_x2,n), where dim_x1 is the number of discrete evaluations of the
surface along dimension one, dim_x2 the same along dimension two, n the number of time instants.}
}
\value{
\strong{\verb{numeric matrix}}, as described above.
}
\description{
Wrap an numeric array into an numeric matrix as suitable input for \link{PPC_KO_2d} and \link{KO_check_hps_2d}. Each column of the output matrix represents a time instants,
while each rows a discrete evaluation of the surface. The grid is stored column-wise.
}
\details{
NaNs values have to be passed anyway as explained in \link{PPC_KO_2d}.
}
\examples{
Xt = array(c(1,2,NaN,3,4,5,NaN,7),dim=c(2,2,2))
PPCKO::data_2d_wrapper_from_array(Xt)
# return  [  1,   4]
#         [  2,   5]
#         [NaN, NaN]
#         [  3,   7]
}
\references{
Source code: \href{https://github.com/AndreaEnricoFranzoni/PPCforAutoregressiveOperator}{PPCKO implementation}
}
\seealso{
\link{PPC_KO_2d}, \link{KO_check_hps_2d}
}
\author{
Andrea Enrico Franzoni
}
