% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC_KO_Rinterface.R
\name{KO_check_hps}
\alias{KO_check_hps}
\title{KO_check_hps}
\arguments{
\item{Xt}{\strong{\verb{numeric matrix}}. Each row (m) represents a point of the domain in which the curve is evaluated.
Each column (n) represents a time instant.}
}
\value{
\strong{\code{list}} whose items are:
\itemize{
\item 'P-values ADF': vector containing the pointwise ADF test p-values.
}
}
\description{
Evaluate the pointwise Augmented DIckey-Fuller (ADF) test p-values for the available evaluations of the curve. Implemented as in \code{tseries} CRAN package, written in C++ for efficiency purposes.
}
\references{
\itemize{
\item Source code: \href{https://github.com/AndreaEnricoFranzoni/PPCforAutoregressiveOperator}{PPCKO implementation}
\item Original R-implementation: \href{https://cran.r-project.org/web/packages/tseries/index.html}{tseries}
}
}
\author{
Andrea Enrico Franzoni
}
